add warehouse new methods and change Integer -> Long products quantity
Changes in API, DTOs and everything related to this. New methods:
POST /api/v1/warehouse/shipped
POST /api/v1/warehouse/return
POST /api/v1/warehouse/assembly

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/interaction/client/feign/shopping/cart/ShoppingCartFallback.java
@@ -20,7 +20,7 @@ public class ShoppingCartFallback implements ShoppingCartClientFeign {
     }
 
     @Override
-    public ShoppingCartDto addProductsToShoppingCart(Map<UUID, Integer> products, String username) {
+    public ShoppingCartDto addProductsToShoppingCart(Map<UUID, Long> products, String username) {
         ShoppingCartFallbackException cause = new ShoppingCartFallbackException();
         log.error(cause.getMessage(), cause);
         throw cause;
